Language.Marlowe.Core.V1.Semantics

Description

Marlowe: financial contracts domain specific language for blockchain

Here we present a reference implementation of Marlowe, domain-specific language targeted at the execution of financial contracts in the style of Peyton Jones et al on Cardano.

This is the Haskell implementation of Marlowe semantics for Cardano.

Semantics

Semantics is based on https://github.com/input-output-hk/marlowe/blob/stable/src/Semantics.hs

Marlowe Contract execution is a chain of transactions, where remaining contract and its state is passed through Datum, and actions (i.e. Choices) are passed as Redeemer Script

Validation Script is always the same Marlowe interpreter implementation, available below.

Semantics

data MarloweData

This data type is a content of a contract's Datum

Constructors

MarloweData

Fields

marloweParams :: MarloweParams

marloweState :: State

marloweContract :: Contract

newtype MarloweParams

Parameters constant during the course of a contract.

Constructors

MarloweParams

Fields

rolesCurrency :: CurrencySymbol

data Payment

Payment occurs during Pay contract evaluation, and when positive balances are payed out on contract closure.

Constructors

Payment AccountId Payee Token Integer

data TransactionInput

Marlowe transaction input.

Constructors

TransactionInput

Fields

txInterval :: TimeInterval

txInputs :: [Input]

data TransactionOutput

Marlowe transaction output.

Constructors

TransactionOutput

Fields

txOutWarnings :: [TransactionWarning]

txOutPayments :: [Payment]

txOutState :: State

txOutContract :: Contract

Error TransactionError

computeTransaction :: TransactionInput -> State -> Contract -> TransactionOutput

Try to compute outputs of a transaction given its inputs, a contract, and it's State

playTrace :: POSIXTime -> Contract -> [TransactionInput] -> TransactionOutput

Run a set of inputs starting from a contract and empty state, reporting the result.

Supporting Functions

addMoneyToAccount :: AccountId -> Token -> Integer -> Accounts -> Accounts

Add the given amount of money to an account (only if it is positive). Return the updated Map.

applyAction :: Environment -> State -> InputContent -> Action -> ApplyAction

Try to apply a single input content to a single action.

applyAllInputs :: Environment -> State -> Contract -> [Input] -> ApplyAllResult

Apply a list of Inputs to the contract.

applyCases :: Environment -> State -> Input -> [Case Contract] -> ApplyResult

Try to apply an input to a list of cases, accepting the first match.

applyInput :: Environment -> State -> Input -> Contract -> ApplyResult

Apply a single Input to a current contract.

convertReduceWarnings :: [ReduceWarning] -> [TransactionWarning]

Propagate ReduceWarning to TransactionWarning.

evalObservation :: Environment -> State -> Observation -> Bool

Evaluate Observation to Bool.

evalValue :: Environment -> State -> Value Observation -> Integer

Evaluates Value given current State and Environment.

fixInterval :: TimeInterval -> State -> IntervalResult

Checks interval and trims it if necessary.

getContinuation :: Input -> Case Contract -> Maybe Contract

Try to get a continuation from a pair of Input and Case.

giveMoney :: AccountId -> Payee -> Token -> Integer -> Accounts -> (ReduceEffect, Accounts)

Gives the given amount of money to the given payee. Returns the appropriate effect and updated accounts.

moneyInAccount :: AccountId -> Token -> Accounts -> Integer

Obtains the amount of money available an account.

playTraceAux :: TransactionOutput -> [TransactionInput] -> TransactionOutput

Run a set of inputs starting from the results of a transaction, reporting the new result.

reduceContractStep :: Environment -> State -> Contract -> ReduceStepResult

Carry a step of the contract with no inputs.

reduceContractUntilQuiescent :: Environment -> State -> Contract -> ReduceResult

Reduce a contract until it cannot be reduced more.

refundOne :: Accounts -> Maybe ((Party, Token, Integer), Accounts)

Pick the first account with money in it.

updateMoneyInAccount :: AccountId -> Token -> Integer -> Accounts -> Accounts

Sets the amount of money available in an account.

Supporting Types

data ApplyAction

Result of applying an action to a contract.

Constructors

AppliedAction ApplyWarning State

NotAppliedAction

data ApplyAllResult

Result of applyAllInputs

Constructors

ApplyAllSuccess Bool [TransactionWarning] [Payment] State Contract

ApplyAllNoMatchError

ApplyAllAmbiguousTimeIntervalError

ApplyAllHashMismatch

data ApplyResult

Result of applyCases

Constructors

Applied ApplyWarning State Contract

ApplyNoMatchError

ApplyHashMismatch

data ApplyWarning

Warning of applyCases

Constructors

ApplyNoWarning

ApplyNonPositiveDeposit Party AccountId Token Integer

data ReduceEffect

Effect of reduceContractStep computation

Constructors

ReduceWithPayment Payment

ReduceNoPayment

data ReduceResult

Result of reduceContractUntilQuiescent

Constructors

ContractQuiescent Bool [ReduceWarning] [Payment] State Contract

RRAmbiguousTimeIntervalError

data ReduceStepResult

Result of reduceContractStep

Constructors

Reduced ReduceWarning ReduceEffect State Contract

NotReduced

AmbiguousTimeIntervalReductionError

data ReduceWarning

Warning during reduceContractStep

Constructors

ReduceNoWarning

ReduceNonPositivePay AccountId Payee Token Integer

ReducePartialPay AccountId Payee Token Integer Integer

ReduceShadowing ValueId Integer Integer

ReduceAssertionFailed

data TransactionError

Transaction error

Constructors

TEAmbiguousTimeIntervalError

TEApplyNoMatchError

TEIntervalError IntervalError

TEUselessTransaction

TEHashMismatch

data TransactionWarning

Warnings during transaction computation

Constructors

TransactionNonPositiveDeposit Party AccountId Token Integer

TransactionNonPositivePay AccountId Payee Token Integer

TransactionPartialPay AccountId Payee Token Integer Integer

TransactionShadowing ValueId Integer Integer

TransactionAssertionFailed

Utility Functions

allBalancesArePositive :: State -> Bool

Check that all accounts have positive balance.

contractLifespanUpperBound :: Contract -> POSIXTime

Calculates an upper bound for the maximum lifespan of a contract (assuming is not merkleized)

isClose :: Contract -> Bool

Check if a contract is just Close.

notClose :: Contract -> Bool

Check if a contract is not just Close.

paymentMoney :: Payment -> Money

Extract the money value from a payment.

totalBalance :: Accounts -> Money

Total the balance in all accounts.

Orphan instances

Lift DefaultUni IntervalError

Instance details

Methods

lift :: IntervalError -> RTCompile DefaultUni fun (Term TyName Name DefaultUni fun ())

Lift DefaultUni IntervalResult

Instance details

Methods

lift :: IntervalResult -> RTCompile DefaultUni fun (Term TyName Name DefaultUni fun ())

Typeable DefaultUni IntervalError

Instance details

Methods

typeRep :: Proxy IntervalError -> RTCompile DefaultUni fun (Type TyName DefaultUni ())

Typeable DefaultUni IntervalResult

Instance details

Methods

typeRep :: Proxy IntervalResult -> RTCompile DefaultUni fun (Type TyName DefaultUni ())