Language.Marlowe.Core.V1.Semantics.Types

Description

Types for Marlowe semantics

Type Aliases

type AccountId = Party

A party's internal account in a contract.

type Accounts = Map (AccountId, Token) Integer

The accounts in a contract.

type ChoiceName = BuiltinByteString

The name of a choice in a contract.

type ChosenNum = Integer

A numeric choice in a contract.

type Money = Value

A multi-asset value.

type TimeInterval = (POSIXTime, POSIXTime)

The time validity range for a Marlowe transaction, inclusive of both endpoints.

type Timeout = POSIXTime

A timeout in a contract.

Contract Types

data Action

Actions happen at particular points during execution. Three kinds of action are possible:

A Deposit n p v makes a deposit of value v into account n belonging to party p.

A choice is made for a particular id with a list of bounds on the values that are acceptable. For example, [(0, 0), (3, 5] offers the choice of one of 0, 3, 4 and 5.

The contract is notified that a particular observation be made. Typically this would be done by one of the parties, or one of their wallets acting automatically.

Constructors

Deposit AccountId Party Token (Value Observation)

Choice ChoiceId [Bound]

Notify Observation

data Bound

The (inclusive) bound on a choice number.

Constructors

Bound Integer Integer

data Case a

A case is a branch of a when clause, guarded by an action. The continuation of the contrack may be merkleized or not.

Plutus doesn't support mutually recursive data types yet. datatype Case is mutually recurvive with Contract

Constructors

Case Action a

MerkleizedCase Action BuiltinByteString

data ChoiceId

Choices – of integers – are identified by ChoiceId which combines a name for the choice with the Party who had made the choice.

Constructors

ChoiceId BuiltinByteString Party

data Contract

Marlowe has six ways of building contracts. Five of these – Pay, Let, If, When and Assert – build a complex contract from simpler contracts, and the sixth, Close, is a simple contract. At each step of execution, as well as returning a new state and continuation contract, it is possible that effects – payments – and warnings can be generated too.

Constructors

Close

Pay AccountId Payee Token (Value Observation) Contract

If Observation Contract Contract

When [Case Contract] Timeout Contract

Let ValueId (Value Observation) Contract

Assert Observation Contract

newtype Environment

Execution environment. Contains a time interval of a transaction.

Constructors

Environment

Fields

timeInterval :: TimeInterval

data Input

Input to a contract, which may include the merkleized continuation of the contract and its hash.

Constructors

NormalInput InputContent

MerkleizedInput InputContent BuiltinByteString Contract

data InputContent

Input for a Marlowe contract. Correspond to expected Actions.

Constructors

IDeposit AccountId Party Token Integer

IChoice ChoiceId ChosenNum

INotify

data IntervalResult

Result of fixInterval

Constructors

IntervalTrimmed Environment State

IntervalError IntervalError

data Observation

Observations are Boolean values derived by comparing values, and can be combined using the standard Boolean operators.

It is also possible to observe whether any choice has been made (for a particular identified choice).

Constructors

AndObs Observation Observation

OrObs Observation Observation

NotObs Observation

ChoseSomething ChoiceId

ValueGE (Value Observation) (Value Observation)

ValueGT (Value Observation) (Value Observation)

ValueLT (Value Observation) (Value Observation)

ValueLE (Value Observation) (Value Observation)

ValueEQ (Value Observation) (Value Observation)

TrueObs

FalseObs

data Party

A Party to a contractt.

Constructors

Address Network Address
Party identified by a network address.

Role TokenName
Party identified by a role token name.

data Payee

A payment can be made to one of the parties to the contract, or to one of the accounts of the contract, and this is reflected in the definition.

Constructors

Account AccountId

Party Party

data State

Marlowe contract internal state. Stored in a Datum of a transaction output.

Constructors

State

Fields

accounts :: Accounts

choices :: Map ChoiceId ChosenNum

boundValues :: Map ValueId Integer

minTime :: POSIXTime

data Token

Token - represents a currency or token, it groups a pair of a currency symbol and token name.

Constructors

Token CurrencySymbol TokenName

data Value a

Values include some quantities that change with time, including “the time interval”, “the current balance of an account”, and any choices that have already been made.

Values can also be scaled, and combined using addition, subtraction, and negation.

Constructors

AvailableMoney AccountId Token

Constant Integer

NegValue (Value a)

AddValue (Value a) (Value a)

SubValue (Value a) (Value a)

MulValue (Value a) (Value a)

DivValue (Value a) (Value a)

ChoiceValue ChoiceId

TimeIntervalStart

TimeIntervalEnd

UseValue ValueId

Cond a (Value a) (Value a)

newtype ValueId

Values, as defined using Let ar e identified by name, and can be used by UseValue construct.

Constructors

ValueId BuiltinByteString

Error Types

data IntervalError

Time interval errors. InvalidInterval means slotStart > slotEnd, and IntervalInPastError means time interval is in the past, relative to the contract.

These errors should never occur, but we are always prepared.

Constructors

InvalidInterval TimeInterval

IntervalInPastError POSIXTime TimeInterval

Utility Functions

emptyState :: POSIXTime -> State

Empty State for a given minimal POSIXTime

getAction :: Case a -> Action

Extract the Action from a Case.

getInputContent :: Input -> InputContent

Extract the content of input.

inBounds :: ChosenNum -> [Bound] -> Bool

Check if a num is withint a list of inclusive bounds.

Orphan instances

Data Address

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> Address -> c Address

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c Address

toConstr :: Address -> Constr

dataTypeOf :: Address -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c Address)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c Address)

gmapT :: (forall b. Data b => b -> b) -> Address -> Address

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> Address -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> Address -> r

gmapQ :: (forall d. Data d => d -> u) -> Address -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> Address -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> Address -> m Address

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> Address -> m Address

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> Address -> m Address

Data Credential

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> Credential -> c Credential

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c Credential

toConstr :: Credential -> Constr

dataTypeOf :: Credential -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c Credential)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c Credential)

gmapT :: (forall b. Data b => b -> b) -> Credential -> Credential

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> Credential -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> Credential -> r

gmapQ :: (forall d. Data d => d -> u) -> Credential -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> Credential -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> Credential -> m Credential

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> Credential -> m Credential

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> Credential -> m Credential

Data StakingCredential

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> StakingCredential -> c StakingCredential

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c StakingCredential

toConstr :: StakingCredential -> Constr

dataTypeOf :: StakingCredential -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c StakingCredential)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c StakingCredential)

gmapT :: (forall b. Data b => b -> b) -> StakingCredential -> StakingCredential

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> StakingCredential -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> StakingCredential -> r

gmapQ :: (forall d. Data d => d -> u) -> StakingCredential -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> StakingCredential -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> StakingCredential -> m StakingCredential

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> StakingCredential -> m StakingCredential

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> StakingCredential -> m StakingCredential

Data PubKeyHash

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> PubKeyHash -> c PubKeyHash

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c PubKeyHash

toConstr :: PubKeyHash -> Constr

dataTypeOf :: PubKeyHash -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c PubKeyHash)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c PubKeyHash)

gmapT :: (forall b. Data b => b -> b) -> PubKeyHash -> PubKeyHash

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> PubKeyHash -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> PubKeyHash -> r

gmapQ :: (forall d. Data d => d -> u) -> PubKeyHash -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> PubKeyHash -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> PubKeyHash -> m PubKeyHash

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> PubKeyHash -> m PubKeyHash

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> PubKeyHash -> m PubKeyHash

Data ScriptHash

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> ScriptHash -> c ScriptHash

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c ScriptHash

toConstr :: ScriptHash -> Constr

dataTypeOf :: ScriptHash -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c ScriptHash)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c ScriptHash)

gmapT :: (forall b. Data b => b -> b) -> ScriptHash -> ScriptHash

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> ScriptHash -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> ScriptHash -> r

gmapQ :: (forall d. Data d => d -> u) -> ScriptHash -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> ScriptHash -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> ScriptHash -> m ScriptHash

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> ScriptHash -> m ScriptHash

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> ScriptHash -> m ScriptHash

Data POSIXTime

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> POSIXTime -> c POSIXTime

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c POSIXTime

toConstr :: POSIXTime -> Constr

dataTypeOf :: POSIXTime -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c POSIXTime)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c POSIXTime)

gmapT :: (forall b. Data b => b -> b) -> POSIXTime -> POSIXTime

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> POSIXTime -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> POSIXTime -> r

gmapQ :: (forall d. Data d => d -> u) -> POSIXTime -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> POSIXTime -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> POSIXTime -> m POSIXTime

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> POSIXTime -> m POSIXTime

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> POSIXTime -> m POSIXTime