Language.Marlowe.Scripts.Semantics

Description

Marlowe validators.

Types

type MarloweInput = [MarloweTxInput]

Input to a Marlowe transaction.

data MarloweTxInput

A single input applied in the Marlowe semantics validator.

Constructors

Input InputContent

MerkleizedTxInput InputContent BuiltinByteString

Semantics Validator

marloweValidatorHash :: ScriptHash

The hash of the Marlowe semantics validator.

marloweValidatorBytes :: SerialisedScript

The serialisation of the Marlowe semantics validator.

marloweValidator :: CompiledCode (BuiltinData -> BuiltinData -> BuiltinData -> ())

The validator for Marlowe semantics.

mkMarloweValidator

Arguments

:: ScriptHash
The hash of the corresponding Marlowe payout validator.

-> MarloweData
The datum is the Marlowe parameters, state, and contract.

-> MarloweInput
The redeemer is the list of inputs applied to the contract.

-> ScriptContext
The script context.

-> Bool
Whether the transaction validated.

The Marlowe semantics validator.

Utilities

marloweTxInputsFromInputs :: [Input] -> [MarloweTxInput]

Convert semantics inputs to transaction inputs.